# docs/swagger.yaml
basePath: /
definitions:
  domain.Order:
    properties:
      createdAt:
        type: string
      customerID:
        type: integer
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/domain.OrderItem'
        type: array
      status:
        type: string
      total:
        type: integer
      updatedAt:
        type: string
    type: object
  domain.OrderItem:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      orderID:
        type: integer
      productID:
        type: integer
      quantity:
        type: integer
      unitPrice:
        type: integer
    type: object
  dto.CreateOrderItemRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.CreateOrderRequest:
    properties:
      customer_id:
        type: integer
      items:
        items:
          $ref: '#/definitions/dto.CreateOrderItemRequest'
        minItems: 1
        type: array
    required:
    - customer_id
    - items
    type: object
  dto.CreateProductRequest:
    properties:
      code:
        maxLength: 64
        type: string
      name:
        maxLength: 255
        type: string
      price:
        minimum: 0
        type: integer
      stock:
        minimum: 0
        type: integer
    required:
    - code
    - name
    type: object
  dto.ProductResponse:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
  dto.UpdateProductRequest:
    properties:
      code:
        maxLength: 64
        type: string
      name:
        maxLength: 255
        type: string
      price:
        minimum: 0
        type: integer
      stock:
        minimum: 0
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  title: API Hệ thống tích hợp Postgres, ES, Redis
  version: "1.0"
paths:
  /orders:
    post:
      consumes:
      - application/json
      description: Creates an order and publishes an OrderCreated event.
      parameters:
      - description: Order payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an order
  /products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a product
  /products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a product
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a product
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a product
  /search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search products
swagger: "2.0"